api_consumo/requirements.txt
django==4.1.3
djangorestframework==3.14.0
django-cors-headers==3.13.0
requests==2.28.1